requirements.txt
pandas>=1.0.0
biopython>=1.70